requirements.txt
# === Core AI & Automation ===
openai>=1.44.0
python-dotenv>=1.0.1
requests>=2.32.3

# === Text-to-Speech / Voice ===
elevenlabs>=1.3.0

# === Video Generation & Editing ===
moviepy==2.0.0
imageio[ffmpeg]>=2.37.0
pillow>=10.2.0,<11.0
numpy>=2.0.0
decorator>=5.2.1
proglog>=0.1.12

# === Google YouTube API ===
google-auth>=2.34.0
google-auth-oauthlib>=1.2.1
google-api-python-client>=2.151.0

# === Logging & Utils ===
tqdm>=4.66.5
colorama>=0.4.6

# === Compatibility Fixes ===
urllib3==1.26.20
importlib_metadata==6.8.0
